% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesselle-package.R
\docType{package}
\name{tesselle-package}
\alias{tesselle-package}
\alias{_PACKAGE}
\alias{tesselle}
\title{tesselle: Easily Install and Load 'tesselle' Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Easy install and load key packages from the 'tesselle' suite in a single step. The 'tesselle' suite is a collection of packages for research and teaching in archaeology. These packages focus on quantitative analysis methods developed for archaeology. The 'tesselle' packages are designed to work seamlessly together and to complement general-purpose and other specialized statistical packages. These packages can be used to explore and analyze common data types in archaeology: count data, compositional data and chronological data. Learn more about 'tesselle' at \url{https://www.tesselle.org}.
}
\details{
\tabular{ll}{
\strong{Package:} \tab tesselle \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.2.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tesselle/tesselle}
  \item \url{https://packages.tesselle.org/tesselle/}
  \item Report bugs at \url{https://github.com/tesselle/tesselle/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order):

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
